chore: copilot adapters, test updates, docs restructure, coverage config
- qyl.copilot: InstrumentedChatClient, InstrumentedAIFunction, ChatClientExtensions adapters
- qyl.copilot: GitHubPkceFlow auth
- tests: HealthEndpointTests + HealthUiServiceTests updated; TrxReport + CodeCoverage packages added
- docs: remove stale ADRs/plans (superseded); add decisions/, roadmap/, screenshots/
- .codecov.yml: coverage config
- CHANGELOG.md, .gitignore, qyl.slnx updated

+++ src/qyl.collector/Storage/DuckDbStore.Services.cs
@@ -348,8 +348,8 @@ public sealed partial class DuckDbStore
                     s.service_name,
                     'unknown' AS service_instance_id,
                     'traditional' AS service_type,
-                    make_timestamp(MIN(s.start_time_unix_nano) / 1000) AS first_seen,
-                    make_timestamp(MAX(s.start_time_unix_nano) / 1000) AS last_seen,
+                    make_timestamp((MIN(s.start_time_unix_nano) / 1000)::BIGINT) AS first_seen,
+                    make_timestamp((MAX(s.start_time_unix_nano) / 1000)::BIGINT) AS last_seen,
                     'active' AS status
                 FROM spans s
                 WHERE s.service_name IS NOT NULL
